LayoutTests/http/tests/navigation/document-open-delayed-adds-history-item.html

<html>
<head>
<script src="resources/document-open.js"></script>
<script>
    function runTest()
    {
        window.firstVisit = true;
        document.open();
        document.writeln("<script>" + window.stopTest + "<" + "/script>");
        document.writeln("<body onload='stopTest();'>");
        document.writeln("This generated document and its contents should be in the back/forward list...");
        setTimeout("runTest2();", 0);
    }
    function runTest2()
    {
        document.writeln("even with a delayed close.");
        document.writeln("</body>");
        document.close();
    }
</script>
</head>
<body onload="startTest()">
This tests that document.open creates a back/forward item, and the page loads
correctly even when writes are split across a timeout.
</body>
</html>
