LayoutTests/http/tests/navigation/document-open-delayed-adds-history-item-expected.txt

This generated document and its contents should be in the back/forward list... even with a delayed close.

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/document-open-delayed-adds-history-item.html  **nav target**
curr->  http://127.0.0.1:8000/navigation/document-open-delayed-adds-history-item.html  **nav target**
        http://127.0.0.1:8000/navigation/resources/document-open-page-2.html  **nav target**
===============================================
